+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:fuks_app/firebase_options.dart';
+import 'package:fuks_app/ui/pages/event/event_page.dart';
 import 'package:fuks_app/ui/pages/events/events_page.dart';
 import 'package:fuks_app/ui/pages/kt/kt_page.dart';
 import 'package:fuks_app/ui/pages/projects/projects_body.dart';
@@ -72,6 +73,7 @@ class _FuksAppState extends State<FuksApp> {
         SettingsPage.route: (context) => const SettingsPage(),
         HomePage.route: (context) => const HomePage(),
         EventsPage.route: (context) => const EventsPage(),
+        EventPage.route: (context) => const EventPage(),
         ProjectPage.route: (context) => const ProjectPage(),
         ProjectsPage.route: (context) => const ProjectsPage(),
         KTPage.route: (context) => const KTPage(),