+++ lib/ui/pages/office/connection_status.dart
@@ -78,7 +78,7 @@ class _ConnectionStatusState extends State<ConnectionStatus> {
           alignment: WrapAlignment.center,
           children: [
             if (Platform.isAndroid)
-              OutlinedButton.icon(
+              TextButton.icon(
                 onPressed: () {
                   if (Platform.isAndroid) {
                     AppSettings.openAppSettings(
@@ -88,11 +88,8 @@ class _ConnectionStatusState extends State<ConnectionStatus> {
                 },
                 icon: const Icon(Icons.wifi),
                 label: const Text('WLAN Überprüfen'),
-                style: OutlinedButton.styleFrom(
-                  foregroundColor: colorScheme.secondary,
-                ),
               ),
-            FilledButton.icon(
+            FilledButton.tonalIcon(
               onPressed: !_isRefreshing
                   ? () {
                       setState(() => _isRefreshing = true);
@@ -110,15 +107,11 @@ class _ConnectionStatusState extends State<ConnectionStatus> {
                       height: 16,
                       child: CircularProgressIndicator(
                         strokeWidth: 2,
-                        color: colorScheme.onSecondary,
+                        color: colorScheme.onPrimaryContainer,
                       ),
                     )
                   : const Icon(Icons.refresh),
               label: const Text('Wiederholen'),
-              style: TextButton.styleFrom(
-                foregroundColor: colorScheme.onSecondary,
-                backgroundColor: colorScheme.secondary,
-              ),
             ),
           ],
         ),