+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2]
+
+### Fix
+
+- Increase timeout for doorman connection
+
+### Added
+
+- Added fuks colors and fonts
+
+### Changed
+
+- Update home view
+- Update events view
+- Update projects view
+
 ## [1.2.1] - 2023-10-22
 
 ### Changed